port(lhc-rs): wave 6 — thread view (dual-verified)
Implements the thread-view surface and internals, host-agnostic fixture seams, lexical materialization, live abort handling, nested source-state provenance, exhaustive closed vocabularies, and exact integer-domain readers.

Amendment I (forced by TS/runtime evidence): profile lowerBound and percentage chains are f64, selection keeps fractional arithmetic, and all diagnostics/persisted values use the shared Node-compatible JavaScript number lane. The prior >=1e21 full-decimal divergence is superseded for reachable profile values.

Committed oracle: fixtures/profile-number-cases.jsonl, 26 Node-generated cases, SHA-256 abe4c924f3dc91789e75eab2567799c4ad17af6af1a82914b3c30c8de3c19068. Real producers cover resolve, selection, compact, raw SQLite, receipt, describe, inspect boundary, nested derivation counts, fractional truncation, and large exponents.

Implementation: Cursor 20260725-021954-580579; repairs 20260725-025527-b4d1d6, 20260725-032731-7a270a, 20260725-034616-a4ca1a, 20260725-035556-d333ca, and 20260725-042615-c0344f, all on cursor-grok-4.5-high-fast.

Verification: initial Sol 20260725-023658-adae38 and Copilot-Fable 20260725-023658-bedeab failed and drove repair. Copilot-Fable r4 20260725-040456-301c90 passed the full producer/oracle scope. Sol r4 20260725-040456-03703f found two missed rounded-bound readers; Sol r5 20260725-043457-51c591 passed direct boundary/aggregate/prune mutation confirmation. Fable narrow integer-reader overclaim is explicitly overridden in the ledger.

Gate: classified 496 = 169 passed + 312 honest phase-2 todos + 15 TS-mirroring ignores; wrong=0, suspicious=0. Owning census is 102 (100 active + 2 ignored). Amendment C view-boundary count correction is recorded.

Project position: Wave 6 of 7, Phase 2 of 3, approximately unit 14 of 18. Wave 7 must still unlock the SDK/inspect/lifecycle remainder to certify the Rust library; all Phase 3 Grok Build integration remains before the user-facing deliverable.

+++ packages/lhc-rs/src/shared_tech/inspect.rs
@@ -104,10 +104,10 @@ pub struct InspectOverview {
 #[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
 #[serde(rename_all = "camelCase")]
 pub struct ViewContentsMetaConfig {
-    /// Token-count target — unified with ViewProfile / StoredViewConfig → i64.
-    pub lower_bound: i64,
-    /// Insertion-ordered band shares.
-    pub percentages: IndexMap<String, i64>,
+    /// Token-count target — Amendment I: TS `number` (fractional-capable).
+    pub lower_bound: f64,
+    /// Insertion-ordered band shares (Amendment I: fractional-capable).
+    pub percentages: IndexMap<String, f64>,
 }
 
 #[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
